third_party/blink/web_tests/fast/css/hover-update-expected.txt

BUG ID: Bugzilla bug 6821 Fix for 5983 will not always update hover correctly.

To run manually, move your mouse over the black box. You should see a green box. If you see any other color, the test fails.

Only green now
PASS successfullyParsed is true

TEST COMPLETE

Hover color before test: rgb(255, 255, 0)
PASS window.getComputedStyle(document.getElementById('b'), null).backgroundColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById('a'), null).backgroundColor is "rgb(0, 128, 0)"
